Get the current value of a cryptocurrency

Get the current value of cryptocurrency in terms of fiat currencies.

coin_value(crypto = "BTC", fiat)

Arguments

crypto
character, cryptocurrency trading symbol, e.g. BTC.

fiat
character, fiat currency trading symbol, e.g., USD.

Value

a data frame.

Details

The current implementation of this function only accepts codecrypto = "BTC". crypto and fiat are not case sensitive. fiat may be a vector. If missing, all available fiat currency values are returned.

Examples

coin_value()

#> # A tibble: 22 x 5
#>    currency       last        buy       sell symbol
#>       <chr>      <dbl>      <dbl>      <dbl>  <chr>
#>  1      USD    5651.37    5651.84    5650.89      $
#>  2      AUD    7176.13    7176.73    7175.52      $
#>  3      BRL   17941.97   17943.48   17940.46     R$
#>  4      CAD    7048.64    7049.23    7048.04      $
#>  5      CHF    5512.74    5513.21    5512.28    CHF
#>  6      CLP 3533798.53 3534095.55 3533501.52      $
#>  7      CNY   37459.04   37462.19   37455.89      ¥
#>  8      DKK   35543.21   35546.20   35540.23     kr
#>  9      EUR    4789.63    4791.06    4788.19      
#> 10      GBP    4293.74    4294.10    4293.38      £
#> # ... with 12 more rows